Simplify / make explicit eol/position logic
The setPosition method allowed selections to be maintained. This is not
used anywhere and therefore that option was removed. The modify selection
method sets both the position and selection.

The setPosition method contained extra logic to "reset" the
positionEolStart state. Instead, positionEolStart state should be
explicitly reset and passed wherever necessary. This required minor
additional logic in the up/down navigation and selection methods.