Handle hard return input and display
Hard returns handling for input, navigation, selection, and rendering. The
last line in the state is now defined by the flow to use an "EOF" constant
if it falls after a hard return. This is needed so that the line created
by the hard return can exist without having an "end" character (since the
end of the text is the hard return).

In setPosition, do not default positionEolStart to true. Instead, default
to the current state of positionEolStart. It helps to be more explicit on
calls, since the complexity of setting this value increases with the
presence of hard returns.

In the word tokenizer default options, add a newline char as a "word".

Improve up/down navigation and selection to deal properly with hard returns
as well as fix some bugs with various corner cases. _modifySelection now
explicitly takes positionEolStart just like setPosition does.