Cursor slanted when italics are active
Cursor slant was working when positioned at an italic character. However,
this commit modifies the slant acccording to whether italics is an active
attribute or not as well i.e. if turning on italics for future characters,
or turning it off.

Made the activeAttributes a part of the component state so that ReactJS
can control the cursor rendering efficiently.

The RichText class now clones attribute objects for setting attributes, so
that only the API methods can be used to change attributes, not just
changing an object reference (which would not propagate through the
replica). Added tests to verify this.