Better EOF handling
In the RichText replica, add some special handling for EOF. While this is
a bit of a leaky abstraction (since the replica itself does not store an
EOF character), it simplifies the calling code a lot if it does not have to
convert to/from a representation with/without EOFs.

When selecting at the end of the document with a trailing newline, use
EOF as the selectionRightChar. This makes the logic simpler and more
consistent.

Small performance improvement in renderSelection: line advance is already
calculated during flow, make use of this to avoid always recalculating in
selection width measurement.