Move input area with cursor
Move the input area vertical position with the cursor. This prevents artifacts such as the browser
scrolling to the top of the editor when the input area receives focus and it is off-screen. Now if the
browser scrolls to the input area the vertical position is the same as the cursor position.