Support HTML paste
When HTML data is pasted, the event now checks whether the clipboard
contains data of mime type text/html. If so, it parses the HTML data using
the new parser, and calls a new insertCharsBatch API for doing the inserts
of the rich text chunks into the editor.

Also handles the IE clipboard API, with a bunch of special handling for
pasting HTML which cannot be retrieved directly from the clipboard event.
See:

http://stackoverflow.com/a/27279218/430128
https://www.lucidchart.com/techblog/2014/12/02/definitive-guide-copying-pasting-javascript/