Temporarily revert: Eliminate extra render cycles on keypresses
When making (batched) calls to setState, React only *enqueues* state
updates -- it does not actually make the change to the state. With the
current structure of the code, this doesn't work properly as subsequent
code needs the updated state.

The solution to this is likely to move to a Flux-type architecture: move
the replica and associated logic into a store. Since the store now becomes
the "source of truth" all code has access to the correct information, and
the state only needs to be updated once the store knows the underlying
editor action has been completely handled.

This will also facilitate testing and future refactoring of the view
components.