Create EditorActions and EditorStore
EditorActions and EditorStore implement the flux unidirectional data flow.
Simplifies the react components by leaving UI concerns only, while most of
the logic is in the EditorStore.

EditorActions are now accessible directly from TextInput, as well as from
external callers.

Also improves:

* testability: all state change logic is isolated in the store

* performance: batches state setting automatically since the store
does not update the component state until the action is complete (which is
important for actions initiated from mousetrap which are not batched
automatically by React)

* will be useful for future undo/redo support since all actions are handled
in the same way and place

* will be useful for future split of the editor into more granular UI
components