Fix scroll to caret after selection
After a selection was made, the React cursor ref was invalidated and so
the caret position was always at the origin.

Instead of destroying the cursor when a selection is made, now it is just
set to invisible. Thus the caret reference remains valid and the window
scroll works with selection changes.