Let RichText char matches shortcut comparisons
Rather than returning a match count which requires all comparisons to be
run, return a boolean which can shortcut subsequent logic when the first
match is found. Most callers only need this.

Support the single instance requiring a count with a new matchCount method
which operates like the old version.