Flow algorithm line outputs chars with chunk indices
Previously, the flow algorithm was outputting the chunks (which contained
the chars within them).

This made it more difficult to extract the chars for other uses because
once has to loop through the chunks. Now, keep the chars in a separate
array and just index into it for the chunks.

This already simplifies the single use of chunks beyond rendering
(positioning based on mouse clicks).