Massive flow performance improvements
The flow algorithm to calculate line breaks and wrapping was very slow when
the editor contents were longer, because it was reflowing the entire
contents of the editor on each change. Now, the flow algorithm can be
passed the region of the modification that caused the reflow, and the
parts of the content that have not changed can be re-used. This also
reduces GC pressure because line content stays in memory for longer rather
than being recreated constantly.

Add charEq and charArrayEq functions into EditorCommon. Will move charEq
from RichText.js in a subsequent commit.

Simplify Line class to avoid passing in charIds. Now calculate and store
charIds lazily when required.