Support clicking in margins, add top/bottom margins
When clicking or selecting outside the bounds of the text bounding box
but within the editor margins, the cursor now navigates to the correct
position:

top/bottom: maintain x position, set y position to first/last line
left/right: maintain y position, set x position to begin/end of line

Remove the empty Editor component and replace it with the (renamed)
EditorContents component, so that the mouse events can occur on the entire
editor.

Set the top/bottom and left/right margins in the test app.js to different
values so bugs with incorrect margin refs are not hidden.

Remove the temporary red dashed line around line bounding boxes.