Fix TextInput focus regressions and behavior for IE
The IE method to select the clipboard div contents is not usable by other
browsers, and was causing clipboard events to be sent to the html body,
even though the TextArea had the focus. Rename
TextInput._selectNodeContents to TextInput._selectIeClipboardDivContents
and use it only when doing the IE clipboard div focus.

For other browsers, a simple .select() suffices.

Also, do not change the Editor focus state when doing an IE copy/paste
which requires a temporary focus to the IE div.

Lastly, on mouse move over the editor do not focus unless a button is
pressed. Otherwise input may still go to the other location e.g. console.