Editor root component shouldComponentUpdate optimization
The root component checks all state and props in shouldComponentUpdate to determine
whether render should be called. React performance tools were showing a lot of wasted
renders / time in the Editor root.

Unfortunately, this adds some brittleness / maintenance overhead because
shouldComponentUpdate always needs to be updated correctly, otherwise weird bugs will
happen if the editor does not render when it needs to.