Fix remote cursor issue with shouldRenderComponent optimization
Remote cursors were not always being rendered with the
shouldRenderComponent optimization blocks. This is because the remote
cursor state was being mutated directly, causing the shouldRenderComponent
to mistakenly believe the state had not changed.

Modify the remote cursor model to make it simpler to clone and access only
the internal cursor state without dealing with the swarm data. When
modifying remote cursor state, clone the state so that the value in the
state is not mutated.

A better solution is an immutable data structure. TODO later.