Show remote cursors overlapping local cursors
Previously if a remote cursor overlapped a local one, the remote cursor was
suppressed. This is subtly incorrect: if the remote cursor is suppressed
the user may think the other person left. If the local cursor is
suppressed instead, the user knows the other person is still there, and
the lack of their own cursor prompts the local user to click somewhere
else.