Fix cursor name reveal with shouldRenderComponent optimization
Another case of state being mutated directly in the store, causing the
shouldRenderComponent to mistakenly believe the state had not changed.
Create a new instance of the remoteNameReveal structure before mutating
it.

Also modify the remote cursor name reveal to be a Set (to avoid a potential
memory leak with the previous array push/shift mechanism).

A better solution is an immutable Set data structure. TODO later.