Handle IME input properly, key capture only non-char events
IME input is placed into the input area rather than captured by keyboard
events. When the onInput event is triggered, this input is placed into the
editor. It seems to work well for various characters I experimented with
on Linux using the Compose key. More testing is needed.

Capture all ascii characters this same way for consistency and to avoid the
overhead of keycode translation in Javascript that can be handled natively
by the browser.

Fixes #13