Attributes cleanup and state for (future) API
Previously, current attributes were only set at insertion time based on
the selection or insertion point. Or when attribute state was "toggled"
without a selection.

Now, the activeAttributes state is updated when navigation or selections
are made. For mouse-based selections, the attributes are updated when the
mouse-up event occurs (for performance during a selection drag). For
keyboard navigation/selections, or mouse navigation i.e. click, the
attributes are updated immediately by the store.

Now, activeAttributes can be exposed to the API and external state (such
as active attributes on a toolbar) can be updated as navigations and
selections are made.

Lastly, attributes stored into the replica are normalized so that the
minimal normalized object, or null, is stored.

Also fixes #15 and #16.