Refactor ritzy init code, reconfiguration now allowed
The ritzy init code was refactored so that default values are initialized,
and required values checked, once at load time.

Ritzy can now be reconfigured at runtime (after the initial load), and the
Editor component is passed the updated configuration (or props in React
parlance).

Update README to indicate that the ritzy module is optional, and the
Editor component can be used directly in React apps.