API documentation
Add API documentation. The API is documented using JSDoc, and extracted
to an asciidoc file using an extractapi.js script.

Add a preversion script that extracts the API to the docs, as well as
creates a markdown version of the README, which certain tools such as
esdoc, and remote sites such as npm.org, expect. Do not check this into
source control.

Fix an indentation error in the webpack config.