Add box-sizing: content-box at Editor top-level div
The Editor adjusts its internal sizing based on the passed in width
and margin props. In order to do this properly, the top-level div
must have the box sizing model set to content-box, otherwise the
padding (which is used internally for the 'margin' setting) will not
be included in the width if the hosting page uses border-box box
sizing.