Correct the deleted id storage in the RichText CT
Deleted ids were not being updated correctly in the RichText causal
tree. The deleted ids were not being pushed to the correct character.
The tests were passing because Chai does not understand ES6 sets and
was reporting equality despite the Sets being unequal.

Correct the data structure, and update the tests to work around the
Chai limitation.

Fixes #21