Add text insertion @ scheduled time in DebugEditor
For debugging insert concurrency issues, add the ability to schedule an
insert for a specific time (minutes and seconds). Doing this on two or
three or more browsers at the same time allows for testing concurrent
inserts.

NOTE: at this commit, this is failing (issue #22) and can now be
reproduced reliably.