Fix bugs in RichText CRDT for concurrent inserts
Two bugs in the RichText CRDT for concurrent inserts are now fixed:

1) The insertion point for a concurrent insert is modified based on the
length of the other text concurrently inserted, and the text is actually
inserted rather than being ignored (this bug was introduced when
converting from the upstream Swarm Text.js).

2) The ins object passed to the `insert` op method was being modified by
reference, causing the host to send the modified op to peers. Remove
this modification. This modification breaks replay on other peers
because it does not maintain causality in the tree (this bug was
inherited from the upstream Text.js).

Lastly, add a warning for insert keys that are not matched. This should
not happen for now, but once we add tombstone clearing mechanisms it
may occur and should be handled as a normal situation.

Fixes #22.