Improvements to shared cursor handling
The cursor model was not being updated properly on initial add because
the state was being set before the .init event was received from Swarm.
Set the cursor model state on .init instead.

Inject a callback into the store rather than the model itself. This
encapsulates model modifications in the mixin and simplifies the store.

Add a warning if the cursor model is "disconnected" from Swarm.js
(detect based on the model listener count). This seems to happen on
rare occasions but I don't know how to repro. The warning will at least
print some useful information to the log. In addition, try to recover
by creating and registering a new cursor model.